Renamed MainWindow files to match the class name
