+++ qtEditor/MainWindow.cpp
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "MainWindow.h"
 #include "ui_mainwindow.h"
 #include "OgreWidget.h"
 #include "EngineCore/EngineInterface.h"
