+++ qtEditor/MainWindow.h
